Allow string constants of any size
This commit modifies the lexer to allow string constants of any size.

It also rationalises the way the string representing the token in
subtilis_token_t.buf is zero terminated.  Previously, the code to
zero terminate strings was scattered all over the place and it was
hard to prove that the buffer was zero-terminated in all cases.  This
commit now requires users to call subtilis_token_get_text to retrieve
the token text.  This function ensures that the token buffer is
zero terminated before returning the string.  Providing the token
text is always retrieved via this method its consumers can be sure
they always have a zero-terminated string.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>