Check token size once token has been read
This commit modifies the way tokens are checked to see whether they
are too large, i.e., > 255 characters.  Previously, this check was
made while the token was being parsed.  Now the check is performed
after the entire token as been parsed.  This simplifies the code
and also makes parser recovery easier.

A variant of subtilis_token_get_text that returns an error has also
been added.  This is intended for use inside the lexer.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>