Add the ARM walker
The ARM walker is a convenience function for walking through a
subtilis_arm_program_t.  There's a number of places in the code
where this is currently done or soon will be done so it seemed
sensible to encode the knowledge of how to do this in one place.
As we add more ARM instructions there'll be less code to change.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>