Add ARM instruction encoder
This commit adds an ARM instruction encoder.  The compiler now generates
an absolute ARM binary which should be directly executable on RISCOS.  The
encoder only generates code for the subset of the ARM instructions currently
used by the compiler.  More work will be needed before it can be used for
an assembler.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>