Add ARM unit tests
This commit adds some unit tests for the code that generates ARM
code.  In introduces a pseudo ARM simulator to run the generated
code.  This simulator processes subtilis's internal representation
of an ARM program rather than a real ARM binary.  It also
implements a few ARM SWI calls, enough to get the test cases to run.
Note that this is not intended to be a complete ARM simulator.  It
simulates just what it needs for running the unit tests.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>