Introduce sections
This commit introduces the concept of sections.  Previously, we had
IR and ARM abstractions for programs, but this weren't really programs.
They were actually sections.  The existing data structures and variables
have been renamed as appropriate and an IR program abstraction supporting
multiple sections has been introduced.  We haven't added a program
abstraction for the backend yet.  This will be done in a subsequent
commit.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>