Add support for procedures
This commit adds basic support for procedures.  There are no parameters
yet nor are there any local variables.  You can however call procedures
defined with the same source file.  This is a big change which required
modifications to all aspects of the compiler.  I also needed to modify
the arm test harness to accept encoded machine code as its input
rather than the internal representation of an arm program.  This needed
to be done as linking takes place during the encoding phase, currently.
This required the addition of a basic ARM dissasembler, and probably
should have been done in a separate commit, but hey, this is a hobby
project.

Ultimately, the dissasembler may find use inside the compiler itself
and perhaps in some supporting tools.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>