Introduce a union for the target of branch
The subtilis_arm_br_instr_t structure has two fields to represent
the target of a B instruction.  One is used when the arm code is
being built up by the compiler and refers to an internal label
or another section.  The other is used when a fully encoded B
instruction is dissasembled.  The later is a PC relative offset.
As neither of these fields are used at the same time it made
sense to place them into a union.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>