Add new IR op type for calls
Function and procedure calls were implemented as a normal IR instruction.
However, this instruction could not be easily used to pass parameters.
For this reason a new type of IR op type has been introduced, to all
procedure calls with unlimited number of parameters to be passed.  The old
IR CALL op has been deleted.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>