Support more than 4 parameters
This commit adds support for more than 4 parameters.  The extra
parameters are placed onto the stack and the register allocator
is seeded with some fake virtual registers whose spilt offsets
correspond to the stack offsets of the extra args.  This way
the register allocator takes care of restoring these arguments
to real registers when they're needed.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>