Add backend capabilities
The idea here is that the backend will advertise a set of capabilities
that can be viewed by the frontend.  The capabilities indicate whether
or not certain IR instructions can be mapped succintly do the underlying
CPU.  If they cannot, for example DIV on ARM, the front end will generate
a builtin function call to implement the associated operation rather
than an IR opcode.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>