FPA floating registers should start from 4
This commit fixes a bug in the code generator for FPA code
that was using floating registers 0-3.  This violates an
constraint of the register allocator which says that the
set of fixed and floating registers used by the code
generator cannot overlap.  As the code generator uses registers
0 to 3 for parameter passing and return values, they cannot
be used as floating registers.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>