Update next when spilling
The array which kept track of which registers will be used again
once they've been allocated was not updated when registers were
spilt or spilt registers were loaded.  The result being that the
contents of these arrays were potentially incorrect.  This could
lead to the allocator chosing a sub optimal register to spill.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>