Simplify ARM register namespace
The existing namespace for ARM registers is really confusing as the
virtual register and fixed register namespaces can overlap.  To
distinguish between them we need to have an extra enum.  It would be
simpler if we could just use the register number to determine whether
a register is fixed or virtual.  Now we can.  This commit changes
the register namespace so that anything < 16 for ints and 6 for reals
is fixed and anything greater is virtual.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>