Add support for TIME as a read only variable
This commit adds support for TIME as a read only variable.  It
cannot be written to and this feature will probably never be
implemented.  For the interpreter and the arm_vm TIME returns
the number of centiseconds since December 2018.  This is done
to ensure that it returns a positive value.  This date will need
to be updated every year or so.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>