Subsections are now terminated by conditional branches
Previously, subsections were delimited by unconditional branches
and branch targets.  However, this was leading to inefficient code
in which virtual registers were being restored from the stack at
the start of each subsection, even though there was no guarantee
that those registers would be used during the execution of the
subsection due to the presence of conditional branches within it.
By using conditional branches to delimit the end of a subsection
we avoid this problem.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>