Optimise vreg save code
The number of store commands generated to save virtual registers
when exiting a subsection are reduced by this pull request for the
case where the subsection links to two other subsections.  The
virtual registers that need to be stored before the transitions
to both sub-sections are now only stored once.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>