Fix function return reg alloc issue
The register allocator was not treating the BL instruction correctly,
in that it was not recognizing that this function could change the R0
or F0 register.  This could lead to an assertion failure in the
register allocator, to a value in a virtual register being overwritten
or a return value corrupted.  The call that saves and restores registers
before and after function calls was also incorrectly saving the R0
register in some cases.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>