Use <- to return values from functions instead of =
= cannot be used as it is impossible for the parser to know
whether = is used to return a value from a function or should be
part of a previous expression, e.g.,

DEF FNA
  A = 10 * COS(RAD(60))
=A

In this case the compiler treats A = 10 * COS(RAD(60)) = A as
a boolean expression and then thinks the function is unterminated.
So we use <- instead as this cannot be part of an expression.

DEF FNA
  A = 10 * COS(RAD(60))
<-A

Signed-off-by: Mark Ryan <markusdryan@gmail.com>