Fix the spill_stack
The spill stack was not working as expected.  When loading a spilled
value, the value needed to be inserted back into it's correct location
and spill_top should only be modified if it was the top most offset
that was loaded.  When spilling a register we need to search through
the spill stack for any free offsets before chosing the offset at
spill_top.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>