Useless moves matter when saving registers
There was an optimisation in the distance calculators which ignored
instructions like mov r0, r0.  This was not the correct thing to do
as it led to the code that saves registers before function calls
from considering that registers such as r0 was not used when in
fact it might be used sliently by an SWI.  We could also fix the issue
by updating the integer distance calculators to handle SWIs properly.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>