Remove some more weird optimisations in the register allocator
There was another weird optimisation in the register allocator
in which the allocator was detecting sequences of instructions
were a spilled value was loaded into a register and then moved
directly into another register.  The original move was
essentially illiminated by the register allocator but this sometimes
lead to problems so the code is being removed.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>