Remove lexer and parser support for builtins
Turns out that this isn't really needed.  It most cases we want
to have specific code in the parser to handle each builtin.  The
reason is that there will be different code paths for builtins
with constant only parameters, in which case we may want no function
at all, and builtins with variable parameters.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>