Implement INKEY(var) as a builtin
Previously, INKEY(var) was implemented inline in the calling function
but it was generating too much code to be placed inline so it is now
implemented as a new type of builtin, an IR builtin.  IR builtins are
built in functions of IR code that are generated by the frontend.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>