Remove IR builtins from the builtin mechanism
There's no need to automatically add these builtins when we do a
function call.  We can add them before invoking the function call
which is slightly less weird.  So with IR builtins, we add the
section if it does not already exist and populate the section with
the correct code before calling the builtin.  Ultimately, we will
probably do the same thing with backend builtins.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>