Implement floating point versions of RND
And this has caused a lot of problems and resulted in another issue
with the grammar.  The problem is that the BBC basic version of RND
can return two different types depending on the value (and not the
type) of its argument.  As this information cannot be known at
runtime for variables, we have no choice but to return floating
point numbers, even for the whole number cases, which is weird.
So we have

X% = 10
RND(X%)

which returns a whole floating point number between 1 and 10

and

RND(10) which returns an integer.  For compatability this is probably
the best solution as RND(X%) will automatically get converted to
an int when needed, but it's weird and inefficient.  Long term
this is one of the many things about the grammar that will need to
change.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>