Always save R11 when calling a procedure
The code that determines which registers to save across procedure
calls only considers the registers that are used in the basic
block after the procedure call.  This works fine for virtual
registers as the code in arm_sub_section.c ensures that all
virtual registers are correctly saved if necessary before the
end of the basic block.  Such registers will be used at the
end of the basic block and so the code that preserves registers
across procedure calls correctly saves and restores these registers.
However, the arm_sub_section code doesn't save R11 and so if R11
is not used to save virtual registers after the procedure has been
called, it will not be saved and will be corrupted.  This commit
fixes the issue by always saving R11 along with R14 when procedures
are called.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>