Implement constant pooling
Finally, we can have functions > 1024 bytes.  Pools are inserted
in the ARM code when needed and skipped with a branch instruction.

The support procedures in one of the test cases, fpa_logical, have
been removed as they are no longer needed.  Previously, this test
case would not compile as a single procedure due to the presence of
floating point constants and the lack of constant pooling.  Now we
have constant pooling, we no longer need those extra procedures.
Removing them tests our pooling code.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>