Remove code to zero locals on the stack
It's up to the code that creates these local variables to initialise
this memory correctly.  Currently, the only types that actually uses
local data on the stack are arrays and the code that sets up arrays
ensures that they are initialised correctly.  Anyway, the backend
is really the wrong place to do this sort of initialisation.  We now
have a memset function if we need it.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>