Change level of procedures and functions
There was a weird descrepancy in which top level statements in
the main procedure were at level 0, whereas top level statements
in user defined functions and procedures where set to level 1.
This commit modifies things so that top level statements are
always at level 0 regardless of in which procedure they appear.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>