Promote tmps during local assignment
Temporary variables (currently only arrays) returned from functions
can now be promoted to full local variables if they are assigned directly
to local variables.  This prevents a bunch of copying and relieves
pressure on the cleanup stack.  For example, in this code,

a%() := FNArr%(1)()

the temporary value returned by FNArr%(1) is promoted to a%().  There's
no need to create a new local for a%().

Signed-off-by: Mark Ryan <markusdryan@gmail.com>