Add subtilis_type_if_is_integer and cleanup part of for
We add a new method to subtilis_type_if to determine whether a given type
is an integer or no.  Right now there's only one variable integer type but
there will be more going forward.  Ideally, we shouldn't have to change
any of the parser code as long as that code correctly uses the type_if
interface.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>