Report correct error for bad dim parameter
The compiler was reporting an assertion when the ()s were missing from
an array reference passed as the first parameter to the dim statement.
I've spent time on two occassions debugging this issue only to discover
(and forget) that there was a bug in the basic source code and that
the compiler was just reporting the wrong error.  This should all be
fixed now.  The code for dim when used as an expression has been
simplified.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>