Fix a bug with the register allocator and SWIs
The register allocator was including the registers used by SWI
calls in distance calculations.  This meant that if there was
a lot of register pressure, a register initialised prior to
an SWI call could be used as a spill target for another register
before the call was made, leading to unpredictable results.

SWI calls have now been added to the distance calculations and
the allocator can cope with an SWI call that reads from and
writes to a register.

An optimisation has also been made to avoid spilling an allocated
fixed register when allocating a fixed destination register if
the fixed register is already allocated to itself.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>