Limit distance calculations for virtual registers to blocks
This is a small optimisation that limits the distance calculation for
virtual registers to the current basic block.  We still need to
scan the entire function for fixed registers as these registers are not
currently perserved as we move from one basic block to another, although
perhaps they should be.  As long as there's no branches in code generated
by the statements we should be fine, but it's a bit risky relying on this.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>