Disable error checking for graphics commands by default
The current situation was silly.  The frontend was adding code to
check for errors that the backend wasn't generating.  This commit
adds the appropriate code to the backend for detecting errors
caused by the execution of graphics commands and then disables this
code and the code in the frontend that checks for the errors, by
default.  This is configurable via the setting
ingore_graphics_errors.  I can't imagine why you'd ever want to
enable this on RiscOS, but it's there if you want it.  The end
result is that graphics heavy programs should be a little smaller
now.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>