Garbage collect arrays of strings properly
The cleanup code at the end of functions that contain reference
types now checks for arrays of strings and ensures that all the
strings in an array that is being deallocated are dereffed.
This is achieved by adding an additional entry to each reference
type.  Currently, this is just an integer but ultimately it will
be some sort of function pointer.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>