Only use destructors when needed
The code for checking for and calling destructors is now only generated
at the end of a function if that function contains a reference type
with a destructor.  It's also only generated at the end of compound
statments when deleting a reference type with a destructor.  This makes
the code a fair bit smaller.  There's still some overhead for using
types with destructors, but that's automatic memory management for you.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>