Allow some tests to leak memory
When we forcefully quit a program with end we don't bother to deref
any heap allocated objects.  This causes the unit tests to fail
as, on the ARM builds, we fail if we detect a memory leak.  This
commit allows tests that leak memory as they exit prematurely to
pass.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>