Promote temporary strings returned from functions
This optimisation was already present for arrays but missing for
strings.  Previously, when assinging a local string to the return
value of a function, the string data was copied into a temporary
variable on the stack, and then the contents of this temporary
variable was then copied into a local register, also on the stack.
With this change, we simple promote the temporary variable to a
real local variable, thereby avoiding a whole pile of copying.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>