Add custom implementation of alloc and deref for ARM
Previously, these IR instructions were implemented using the Risc OS
Heap SWIs.  However, these APIs turned out to be very slow and also
unavailable on PiTube direct.  They were also poorly documented and
presumabably weren't non fragmenting.  This new implementation of
alloc is.  However, it has revealed a bug in the register allocator
which causes one of the test cases to fail.  This will be fixed
in a subsequent commit.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>