Fix an issue with the register allocator
The distance calculator for integer stran was checking the
destination before the source operands, which breaks the
register allocator and the other uses of distance calculation,
e.g., preserving registers across function calls and basic
block transitions.  Perhaps the restriction of not using the
same virtual register twice in the same instruction as source
and destination can now be lifted.  I need to check.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>