Add heapfree keyword
This new keyword returns the number of unallocated bytes currently in
the heap, i.e., the amount of memory currently available for strings
and arrays.  It's mainly intended for debugging purposes but it could
be used by an application to check that it has enough memory available
when first invoked.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>