Add compiler setting for detecting memory leaks
This new setting, when enabled, forces the compiler to generate code that
checks the amount of free space in the heap at the start and the end
of the program and to report an error if the two values are not equal.
This is nicer than the previous solution which only worked on the ARM
builds.  It's enabled by default for the interpreter and most of the test
cases.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>