Fix an encoding error
There was a nasty bug in the arm_encoder which occurred when encoding
an LDRC statement.  If the constant loaded by the LDRC statement
caused the constant pool to be flushed, the location of the LDRC
statement was incorrectly recorded.  This lead to the B statement
being updated by the linker, rather than the LDR statement, and
this caused all sorts of problems.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>