Another attempt to fix constant pooling
9fb65bf did not completely fix the issue with constant pooling.
It was still possible for a constant pool to be inserted between
the LDR and ADD instructions that are generated by the lca IR
instruction.  As both of these instructions are PR relative, this
broke things.  The proper fix is to force a constant pool flush
if there's not enough room for 1 constant and 2 instructions.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>