Fix a nasty long standing bug in the register allocator
The distance walkers check source operands before destination operands,
but there was a bug in the utility code used to process op2.  It
was increasing the last_used counter if the register was not used
in op2.  But this meant that if the register was used as the desintation
register, we returned the wrong distance value.  Also it meant that
the botched code in prv_is_reg_used_before that keeps looking for
instructions that modify a register, was skipping one instruction.
If that instruction happened to be the one that wrote to the register
this fact would go unnoticed.  Not updating last_used in
prv_dist_handle_op2 fixes the problem.  I wonder whether we can
now deal with IR instructions that read from and modify the same
virtual register.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>