Fix same source/destination register allocation bug
This bug has been present in the register allocator since the very
beginning and I've just been working around it until now, by ensuring
that the compiler didn't generate instructions that read from and
wrote to the same virtual register.  This was a pain and leads to
larger code, so the time has come to finally fix it.  The issue
was that the distance walkers were being used to determine whether
a virtual register is written to in a block of code, and this is
a task to which they are not suited.  The return incorrect results
if they find an instruction that reads from and writes to the same
register.  The solution is to provide custom walkers for this purpose.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>