Remove unnecessary add/mov combination
The compiler was avoiding generating IR instructions that used the same
virtual register as both a source and a destination.  Instead it was
computing the new value and storing that value in a tempory register
and then generating a mov instruction to store the contents of the
temporary register in the target register.  Now we've finally fixed the
register allocator, this extra move instruction is no longer needed
and this commit removes such instructions.  It doesn't make much difference
to the final binary as the register allocator and peephole optimizer were
removing most of this extra moves, but I guess it cleans up the code
and perhaps speeds things up a little.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>