Permit -2147483648 to be used as an integer constant
Normally, when we process negative integers in the lexer
we generate two different tokens, one for the '-' operator
and one for the unsigned number.  This doesn't work for the
smallest possible negative number as we cannot represent it
as a positive number.  So for this one case we return one
single token with the correct integer value.  This requires
and explicit check in the parser to ensure that a statement
like

a% = a% -2147483648

returns the correct error.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>