Provide better errors for DEFPROC and DEFFN
Subtilis requires a space between DEF FN and DEF PROC.  Previously,
it was generating very confusing errors about missing assignment
operators when it encountered code like DEFFNSum(a%).  Now when
it sees something like this it assumes that the user was trying
to declare a function and outputs a more meaningful error.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>