warnings: Fix more silly gcc-10 false positives
gcc-10 is reporting quite a few warnings claiming that a variable may
be used without being initialized, which isn't the case.  It doesn't
seem to like the fact that Subtilis generally doesn't initialise output
arguments of a function when an error is returned.  So even though the
calling function checks the error and itself returns ensuring that
the unitialised variable isn't accessed, gcc-10 can't seem to detect this.
It don't want to disable this warning as it did actually detect one
case where there was an access of an unitialised variable so best to
silence these warnings by setting the values of the variables to some
default value in the calling function.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>