Fix a bug in arm_core
The code to insert instructions was broken as it performs and
alloc which can invalidate the pointer to the instruction after
which it needs to insert the new instruction.  We fix this by
getting the index of the insertion point instruction before  the
alloc and then refreshing the pointer to it after the alloc.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>