Fix a bug in left$ and right$
There was an ill conceived optimisation in these two functions in which
the code would return the existing string if the program requested
the full string, e.g., left$(a$) where len(a$) = 1.  Unfortunately,
this caused problem as the code was creating space for a temporary
variable at compile time that as never getting initialised and pushed
onto the stack.  Doing this inside a block would result in the
wrong variable getting popped off the stack.

Now we always create and populate the temporary variable even if the
full string is to be returned.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>