Prohibit global variable creation after procedure call
Because, currently, the compiler does not zero initialise global
variables at the start of the program.  Procedures accessing
global variables initialised after they had been called would
find that the global variables are uninitialised, which creates
all sorts of problems if the variable is something like an array.
So we disable this for now, at least until the front end is
re-written, so we have an easy way to zero-initialise global
variables.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>