riscv: Fix the stack
We now preserve and restore live registers before and after function
calls.  This is done using the register allocator and not using
a totally separate mechanism as we have done in ARM.  We also fix
the layout of the stack making sure that the stack and all reals
are 8 byte aligned when written and read.

Signed-off-by: Mark Ryan <markusdryan@gmail.com>